+++ app/src/main/java/com/ekylibre/android/MainActivity.java
@@ -80,7 +80,7 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
     private RecyclerView.Adapter adapter;
     public static List<Interventions> interventionsList = new ArrayList<>();
 
-    private SharedPreferences sharedPreferences;
+    private SharedPreferences prefs;
     private SyncResultReceiver resultReceiver;
 
     public static String FARM_ID;
@@ -98,11 +98,11 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
         Enums.buildEnumsTranslation(this);
 
         // Get shared preferences and set title
-        sharedPreferences = this.getSharedPreferences("prefs", Context.MODE_PRIVATE);
-        setTitle(sharedPreferences.getString("current-farm-name", "No name"));
+        prefs = this.getSharedPreferences("prefs", Context.MODE_PRIVATE);
+        setTitle(prefs.getString("current-farm-name", "No name"));
 
         // Get current farm_id
-        FARM_ID = sharedPreferences.getString("current-farm-id", "");
+        FARM_ID = prefs.getString("current-farm-id", "");
 
         // Get locale one time for the app
         LOCALE = getResources().getConfiguration().locale;
@@ -135,7 +135,6 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         //recyclerView.addItemDecoration(new DividerItemDecoration(this, DividerItemDecoration.VERTICAL));
         adapter = new MainAdapter(this, interventionsList);
-
         adapter.registerAdapterDataObserver(new RecyclerView.AdapterDataObserver() {
             @Override
             public void onChanged() {
@@ -148,7 +147,6 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
                 }
             }
         });
-
         recyclerView.setAdapter(adapter);
 
         // All button events
@@ -166,20 +164,13 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
             toast.show();
         });
 
-        swipeRefreshLayout.setOnRefreshListener(() -> {
-            Intent intent = new Intent(this, SyncService.class);
-            intent.setAction(SyncService.ACTION_SYNC_ALL);
-            intent.putExtra("receiver", resultReceiver);
-            startService(intent);
-        });
+        swipeRefreshLayout.setOnRefreshListener(() -> new StartSync(SyncService.ACTION_SYNC_ALL).execute());
 
-        Intent intent = new Intent(this, SyncService.class);
-        intent.setAction(SyncService.FIRST_TIME_SYNC);
-        intent.putExtra("receiver", resultReceiver);
-        startService(intent);
+        // Run a get Sync on startup
+        new StartSync(SyncService.FIRST_TIME_SYNC).execute();
 
-        //        if (!sharedPreferences.getBoolean("showcase-passed", false)) {
-//            SharedPreferences.Editor editor = sharedPreferences.edit();
+//        if (!prefs.getBoolean("showcase-passed", false)) {
+//            SharedPreferences.Editor editor = prefs.edit();
 //            editor.putBoolean("showcase-passed", true);
 //            editor.apply();
 //            new FancyShowCaseView.Builder(this)
@@ -188,89 +179,46 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
 //                    .disableFocusAnimation().build().show();}
     }
 
-
-    public class RefreshToken extends AsyncTask<Void, Void, Void> {
-
-        RefreshToken() {}
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            try {
-                ProviderInstaller.installIfNeeded(getBaseContext());
-            } catch (GooglePlayServicesRepairableException e) {
-                GoogleApiAvailability.getInstance()
-                        .showErrorNotification(getBaseContext(), e.getConnectionStatusCode());
-            } catch (GooglePlayServicesNotAvailableException e) {
-                if (BuildConfig.DEBUG) Log.e(TAG, "GooglePlayServicesNotAvailableException");
-            }
-        }
-
-        @Override
-        protected Void doInBackground(Void... voids) {
-
-            if (BuildConfig.DEBUG) Log.i(TAG, "Refresh token request...");
-
-            AccessToken token = new AccessToken();
-            token.setAccess_token(sharedPreferences.getString("access_token", ""));
-            token.setRefresh_token(sharedPreferences.getString("refresh_token", ""));
-            token.setToken_type("bearer");
-
-            EkylibreAPI ekylibreAPI = ServiceGenerator.createService(EkylibreAPI.class, token);
-            Call<AccessToken> call = ekylibreAPI.getRefreshAccessToken(App.OAUTH_CLIENT_ID, App.OAUTH_CLIENT_SECRET, token.getRefresh_token(), "refresh_token");
-
-            try {
-                retrofit2.Response<AccessToken> response = call.execute();
-                if (response.isSuccessful()) {
-                    AccessToken responseToken = response.body();
-                    SharedPreferences.Editor editor = sharedPreferences.edit();
-                    editor.putString("access_token", responseToken.getAccess_token());
-                    editor.putString("refresh_token", responseToken.getRefresh_token());
-                    editor.putInt("token_created_at", responseToken.getCreated_at());
-                    editor.apply();
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            return null;
-        }
-    }
+//    private void startSync(String action) {
+//        if (!swipeRefreshLayout.isRefreshing())
+//            swipeRefreshLayout.setRefreshing(true);
+//        if (new StartSync().execute()) {
+//            Intent intent = new Intent(this, SyncService.class);
+//            intent.setAction(action);
+//            intent.putExtra("receiver", resultReceiver);
+//            startService(intent);
+//        } else {
+//            if (swipeRefreshLayout.isRefreshing())
+//                swipeRefreshLayout.setRefreshing(false);
+//        }
+//    }
 
     @Override
     protected void onResume() {
         super.onResume();
 
-        if (BuildConfig.DEBUG) Log.i(TAG, "Resuming MainActivity...");
-
-        // Check the token is not expired and request for new one
-        long tokenTime = (long) sharedPreferences.getInt("token_created_at", 0) * 1000;
-        long now = new Date().getTime();
-
-        Log.i(TAG, "Time diff = " + (now - tokenTime));
-
-        if ( now - tokenTime >= 7000000)  // 7200000
-            new RefreshToken().execute();
-        else
-            if (BuildConfig.DEBUG) Log.i(TAG, "Token is up to date");
-
         try {
-            lastSyncTime = LAST_SYNC.parse(sharedPreferences.getString("last-sync-time", "2018-01-01 12:00"));
+            lastSyncTime = LAST_SYNC.parse(prefs.getString("last-sync-time", "2018-01-01 12:00"));
         } catch (ParseException e) {
             e.printStackTrace();
         }
         // TODO auto sync if lastSyncTime < now - 10min
 
+        // Set Farm name as page title
+        setTitle(prefs.getString("current-farm-name", "Synchronisation..."));
+        FARM_ID = prefs.getString("current-farm-id", "");
+
         // Get list filter and update list
-        // String filter = sharedPreferences.getString("filter", FILTER_ALL_INTERVENTIONS);
-        new UpdateList(getBaseContext(), FILTER_ALL_INTERVENTIONS).execute();
+        String filter = prefs.getString("filter", FILTER_ALL_INTERVENTIONS);
+        new UpdateList(this, filter).execute();
+
     }
 
     @Override
     public void onBackPressed() {
         if (!deployMenu(false))
             super.onBackPressed();
-            AppDatabase.revokeInstance();
+        //AppDatabase.revokeInstance();
     }
 
     @Override
@@ -286,7 +234,7 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
         // Handle item selection
         switch (item.getItemId()) {
             case R.id.action_logout:
-                SharedPreferences.Editor editor = sharedPreferences.edit();
+                SharedPreferences.Editor editor = prefs.edit();
                 editor.remove("is_authenticated");
                 editor.remove("access_token");
                 editor.remove("refresh_token");
@@ -309,9 +257,9 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
         swipeRefreshLayout.setRefreshing(false);
 
         if (resultCode == SyncService.DONE) {
-            if (BuildConfig.DEBUG) Log.i(TAG, "Synchronization done");
+            if (BuildConfig.DEBUG) Log.e(TAG, "Synchronization done");
             lastSyncTime = new Date();
-            sharedPreferences.edit().putString("last-sync-time", LAST_SYNC.format(lastSyncTime)).apply();
+            prefs.edit().putString("last-sync-time", LAST_SYNC.format(lastSyncTime)).apply();
             new UpdateList(this, FILTER_ALL_INTERVENTIONS).execute();
         } else if (resultCode == SyncService.FAILED) { //R.string.sync_failure
             String message = bundle.getString("message");
@@ -338,7 +286,7 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
 
         @Override
         protected Void doInBackground(Void... voids) {
-            if (BuildConfig.DEBUG) Log.i(TAG, "Updating interventions recyclerView...");
+            if (BuildConfig.DEBUG) Log.e(TAG, "Updating interventions recyclerView...");
             AppDatabase database = AppDatabase.getInstance(context);
             interventionsList.clear();
             switch (filter) {
@@ -362,7 +310,7 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
     }
 
     private void onInterventionTypeSelected(int type) {
-        if (sharedPreferences.getBoolean("no-crop", true)) {
+        if (prefs.getBoolean("no-crop", true)) {
             Toast toast = Toast.makeText(this, "Vous n'avez aucune parcelle assolée. Rendez-vous en ligne pour commencer !", Toast.LENGTH_LONG);
             toast.setGravity(Gravity.BOTTOM, 0, 200);
             toast.show();
@@ -420,6 +368,92 @@ public class MainActivity extends AppCompatActivity implements SyncResultReceive
             String quantity_name_only = (unit.surface_factor != 0) ? getString(getResources().getIdentifier(unit.quantity_key_only, "string", getPackageName())) : null;
             unitListString.add(name);
             unit.setName(name, quantity_name_only);
+        }
+    }
+
+    /**
+     * Verify token validity. Ask for new one if expired.
+     */
+
+    public class StartSync extends AsyncTask<Void, Void, Boolean> {
+
+        String action;
+
+        StartSync(String action) {
+            this.action = action;
+        }
+
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+
+            if (!swipeRefreshLayout.isRefreshing())
+                swipeRefreshLayout.setRefreshing(true);
+        }
+
+        @Override
+        protected Boolean doInBackground(Void... voids) {
+
+            long tokenTime = (long) prefs.getInt("token_created_at", 0) * 1000;
+            long now = new Date().getTime();
+
+            if (now - tokenTime >= 7000000) {
+
+                // Handle Handshake Errors
+                try {
+                    ProviderInstaller.installIfNeeded(getBaseContext());
+                } catch (GooglePlayServicesRepairableException e) {
+                    Log.e(TAG, "GooglePlayServicesRepairableException");
+                    GoogleApiAvailability.getInstance().showErrorNotification(getBaseContext(), e.getConnectionStatusCode());
+                } catch (GooglePlayServicesNotAvailableException e) {
+                    Log.e(TAG, "GooglePlayServicesNotAvailableException");
+                }
+
+                if (BuildConfig.DEBUG) Log.i(TAG, "Last Token created " + (new Date().getTime() - tokenTime) / 1000 + " seconds ago");
+
+                AccessToken token = new AccessToken();
+                token.setAccess_token(prefs.getString("access_token", ""));
+                token.setRefresh_token(prefs.getString("refresh_token", ""));
+                token.setToken_type("bearer");
+
+                EkylibreAPI ekylibreAPI = ServiceGenerator.createService(EkylibreAPI.class, token);
+                Call<AccessToken> call = ekylibreAPI.getRefreshAccessToken(App.OAUTH_CLIENT_ID, App.OAUTH_CLIENT_SECRET, token.getRefresh_token(), "refresh_token");
+
+                try {
+                    retrofit2.Response<AccessToken> response = call.execute();
+                    if (response.isSuccessful()) {
+                        AccessToken responseToken = response.body();
+                        SharedPreferences.Editor editor = prefs.edit();
+                        editor.putString("access_token", responseToken.getAccess_token());
+                        editor.putString("refresh_token", responseToken.getRefresh_token());
+                        editor.putInt("token_created_at", responseToken.getCreated_at());
+                        editor.apply();
+                        return true;
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+
+            } else {
+                if (BuildConfig.DEBUG) Log.i(TAG, "Token is up to date");
+                return true;
+            }
+            return false;
+        }
+
+        @Override
+        protected void onPostExecute(Boolean aBoolean) {
+            super.onPostExecute(aBoolean);
+
+            if (aBoolean) {
+                Intent intent = new Intent(getBaseContext(), SyncService.class);
+                intent.setAction(action);
+                intent.putExtra("receiver", resultReceiver);
+                startService(intent);
+            } else {
+                if (swipeRefreshLayout.isRefreshing())
+                    swipeRefreshLayout.setRefreshing(false);
+            }
         }
     }
 }