+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.9.3 bêta - 2018-06/14
+### Fixed
+- Improves Oauth token refresh after long idle
+
+## 0.9.2 bêta - 2018-06/14
+### Fixed
+- Corrects Oauth token refresh
+
+
 ## 0.9.1 bêta - 2018-06-12
 ### Fixed
 - Token refresh